quagga 1.1.0
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@30816 52c4871e-980c-0410-b1e0-e73912ce01f8

Former-commit-id: 33208f2f442f1c6a92409a0d44492445f11f2f0e

+++ src/router/quagga/zebra/rt_netlink.h
@@ -30,10 +30,10 @@
 extern int
 addattr32 (struct nlmsghdr *n, size_t maxlen, int type, int data);
 extern int
-addattr_l (struct nlmsghdr *n, size_t maxlen, int type, void *data, int alen);
+addattr_l (struct nlmsghdr *n, size_t maxlen, int type, void *data, size_t alen);
 
 extern int
-rta_addattr_l (struct rtattr *rta, int maxlen, int type, void *data, int alen);
+rta_addattr_l (struct rtattr *rta, size_t maxlen, int type, void *data, size_t alen);
 
 extern const char *
 nl_msg_type_to_str (uint16_t msg_type);
